+++ CHANGES.rst
@@ -21,6 +21,7 @@ Changelog
 - Optimized item duplication process, remove images, advices and relevant annexes (that are not kept) using `_delObject(suppress_events=True)`
   in `zope.lifecycleevent.ObjectCopiedEvent` `onItemCopied` event handler
 - In `MeetingConfig.getMeetingsAcceptingItems`, extracted computation of catalog query into `MeetingConfig._getMeetingsAcceptingItemsQuery`
+- An item may be taken over by members of the `proposingGroup` when it is decided
 
 4.1.25.1 (2020-06-02)
 ---------------------